+++ libs/vkd3d/utils.c
@@ -868,75 +868,6 @@ HRESULT vkd3d_load_vk_device_procs(struct vkd3d_vk_device_procs *procs,
     return S_OK;
 }
 
-#if HAVE_DECL_PROGRAM_INVOCATION_NAME
-
-bool vkd3d_get_program_name(char program_name[VKD3D_PATH_MAX])
-{
-    char *name, *p, *real_path = NULL;
-
-    if ((name = strrchr(program_invocation_name, '/')))
-    {
-        real_path = realpath("/proc/self/exe", NULL);
-
-        /* Try to strip command line arguments. */
-        if (real_path && (p = strrchr(real_path, '/'))
-                && !strncmp(real_path, program_invocation_name, strlen(real_path)))
-        {
-            name = p;
-        }
-
-        ++name;
-    }
-    else if ((name = strrchr(program_invocation_name, '\\')))
-    {
-        ++name;
-    }
-    else
-    {
-        name = program_invocation_name;
-    }
-
-    strncpy(program_name, name, VKD3D_PATH_MAX);
-    program_name[VKD3D_PATH_MAX - 1] = '\0';
-    free(real_path);
-    return true;
-}
-
-#elif defined(_WIN32)
-
-bool vkd3d_get_program_name(char program_name[VKD3D_PATH_MAX])
-{
-    char *name;
-    char exe_path[VKD3D_PATH_MAX];
-    GetModuleFileNameA(NULL, exe_path, VKD3D_PATH_MAX);
-
-    if ((name = strrchr(exe_path, '/')))
-    {
-        ++name;
-    }
-    else if ((name = strrchr(exe_path, '\\')))
-    {
-        ++name;
-    }
-    else
-    {
-        name = exe_path;
-    }
-
-    strncpy(program_name, name, VKD3D_PATH_MAX);
-    return true;
-}
-
-#else
-
-bool vkd3d_get_program_name(char program_name[VKD3D_PATH_MAX])
-{
-    *program_name = '\0';
-    return false;
-}
-
-#endif  /* HAVE_DECL_PROGRAM_INVOCATION_NAME */
-
 static struct vkd3d_private_data *vkd3d_private_store_get_private_data(
         const struct vkd3d_private_store *store, const GUID *tag)
 {
